Conflicts:
	share/data/pr2_model/pr2_left_arm.ors
	share/examples/Motion/RRT-PR2/main.cpp
	share/src/Core/array_typemaps.i
	share/src/Motion/Makefile
	share/src/Motion/motion.cpp
	share/src/Motion/rrt_planner.cpp
	share/src/Motion/rrt_planner.h
	share/src/Motion/taskMap_proxy.h
	share/src/Ors/ors_physx.cpp
	share/src/relational/robotManipulationSimulator.cpp.orig

+++ src/Perception/video_opencv.cpp
@@ -17,7 +17,7 @@
     -----------------------------------------------------------------  */
 
 
-#include "video.h"
+#include "videoEncoder.h"
 #ifdef MT_OPENCV
 
 #undef COUNT
@@ -65,7 +65,7 @@ void VideoEncoder_OpenCV::close(){
 #else //MT_OPENCV
 
 #include <Core/util.h>
-  VideoEncoder_OpenCV::VideoEncoder_OpenCV(const char* filename,double fps){ MT_MSG("WARNING - using dummy Revel module"); };
+  VideoEncoder_OpenCV::VideoEncoder_OpenCV(const char* filename, uint fps){ MT_MSG("WARNING - using dummy Revel module"); };
   void VideoEncoder_OpenCV::addFrame(const byteA& img){};
   void VideoEncoder_OpenCV::close(){};
